pyproject.toml
[project]
name = "ann"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Tanuj-a-bit", email = "tanujs2612@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "python-levenshtein>=0.27.3",
    "scikit-learn>=1.8.0",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "google-generativeai",
    "python-dotenv",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
